chore: bump telescope and treesitter version
Treesitter uses a new configuration/setup process. This means moving
alot of the configs to the "opts" argument. The process will call
treesitter.setup() with these opts by default.

Telescope also is moved to the new version and for some reason on linux
the original setup I had use with declaring local vars outside of the
setup wasn't working. Not sure if this is correct but whatever its
working.