fix: configure lsp support
I originally forked this config from Primeagen. Mason-lsp has been
updated since then and it wasn't working on my machine so I updated to
the latest version (the owner of the repo also changed).

The config also used a different completion plugin that required a ton
of dependencies. I've switched it over to blink since it uses a rust
fuzzy finder. Alot of the key mappings are also removed with this
change.